Create MapEditor.map
File that is read by the loadmap method to set the map layout. The first two numbers indicate the size of the map, the next group of 3 digit numbers indicate the location of the image to be printed into that spot, and the letter indicates which image to use (w being the water image file and g being the grass image file).